Cask & .cask initialization
I want to use Cask to manage my packages. Not sure whether or not I am
supposed to download every single file in the .cask folder, or if I am
supposed to use it as a tool on the other machines I work on, and
simply install and use cask there. I decided to keep every package in
the repo because I may not have access to install cask on the machine
I want to work on, and this way, any reference to packages in the
.cask folder will not throw errors when emacs starts up.