Machine-Specific Inventory
Turns out git will not recognize an empty directory (the directory I
wanted to contain machine-specific stuff also won't have anything in
it pushed to vc). However, because it's so lacking, it made sense to
keep an inventory of what's in each file so if any items were
unaccounted for, I could find them and put them in their place.