Remove org-log-done value as 'note, add function to toggle it
When org-log-done was set to 'note the agenda wouldn't also clock
time. This didn't allow any habits to be tracked, so the function
'toglog toggles 'org-log-done between 'note and 'time, so if I'm
markig habits done, I can toggle 'time, and if I'm working on tasks, I
can toggle 'time.